feat(integracion): Se integraron tokens de firebase en la base de datos

+++ backend/src/main/java/com/t4kash/api/marketplace/entity/Usuario.java
@@ -41,7 +41,7 @@ public class Usuario {
     @Column(name = "id_universidad")
     private Integer idUniversidad;
 
-    public Integer getIdUsuario() {
+    public Long getIdUsuario() {
         return idUsuario;
     }
